backend/requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
azure-cosmos
python-dotenv
tqdm
numpy
